+++ Zero-K.info/Views/Shared/UserDetail.cshtml
@@ -340,7 +340,7 @@
             {
                 <a href="@Url.Action("ResetDB", "Admin")" class="js_confirm">Reset database (copy from live)</a><br/>
             }
-            <a href="@Url.Action("Engines", "Admin")">Engines</a><br />
+            <a href="@Url.Action("Index", "Engines")">Engines</a><br />
             <a href="@Url.Action("ChatHistory", "Lobby")">Lobby chat history</a><br/>
             <a href="@Url.Action("TraceLogs", "Admin")">Infrastructure trace logs</a><br/>
             <a href="@Url.Content("~/ResourceList.aspx")">Resource list</a><br />
